requirements.txt
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
transformers==4.44.0
ultralytics==8.2.78
pillow==10.4.0
pandas==2.2.2
numpy==1.26.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
webcolors==1.11.1
pytesseract==0.3.13
matplotlib==3.9.2
huggingface-hub==0.24.5
colorthief==0.2.1
streamlit==1.37.1